+++ DentalCare/Controllers/TechniqueController.cs
@@ -21,6 +21,13 @@ namespace DentalCare.Controllers
         [Route("technique")]
         public IActionResult Index(int? page, string sortColumn, string sortDirection, string searchQuery)
         {
+            var userRole = HttpContext.Session.GetString("UserRole");
+
+            if (userRole.Contains("D"))
+            {
+                return NotFound();
+            }
+
             ViewBag.Techs = _techniqueService.GetAll();
             var pageNumber = (page ?? 1);
             var pageSize = 10;
@@ -56,6 +63,13 @@ namespace DentalCare.Controllers
         [HttpGet]
         public IActionResult AddTech()
         {
+            var userRole = HttpContext.Session.GetString("UserRole");
+
+            if (userRole.Contains("D"))
+            {
+                return NotFound();
+            }
+
             return View();
         }
 
@@ -70,6 +84,13 @@ namespace DentalCare.Controllers
         [HttpGet]
         public IActionResult Add()
         {
+            var userRole = HttpContext.Session.GetString("UserRole");
+
+            if (userRole.Contains("D"))
+            {
+                return NotFound();
+            }
+
             ViewBag.Techs = _techniqueService.GetAll();
             return View();
         }
@@ -87,6 +108,13 @@ namespace DentalCare.Controllers
         [HttpGet]
         public IActionResult Edit(string id)
         {
+            var userRole = HttpContext.Session.GetString("UserRole");
+
+            if (userRole.Contains("D"))
+            {
+                return NotFound();
+            }
+
             var tech = _techWorkService.Get(id);
             ViewBag.Techs = _techniqueService.GetAll();
             return View(tech);
@@ -103,6 +131,13 @@ namespace DentalCare.Controllers
 
         public IActionResult Delete(string id)
         {
+            var userRole = HttpContext.Session.GetString("UserRole");
+
+            if (userRole.Contains("D"))
+            {
+                return NotFound();
+            }
+
             _techWorkService.Delete(id);
             return RedirectToAction("Index");
         }
